requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
cryptography==43.0.3
dj-database-url==3.0.1
Django==6.0
django-allauth==0.57.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.0.0
PyJWT==2.8.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.4
urllib3==2.6.2
whitenoise==6.11.0
